Changed topic to transgender and added articles on the topic from NewsAPI
Pivoting to a project on the topic of transgender individuals in the news, for the sake of project criteria (has to be a debate of some sort). I also have to use the NewsAPI as a source and at least one other API and one web scraper. I have the compiled articles from NewsAPI and the beginnings of a web scraper for Ground News which has lots of articles and the attached biases. I could perhaps inspect the element of the page directly and borrow that html for my web scraping. This is due to the fact that Im relying on my personal subscription to Ground News to see these articles so I may not be able to access them if I try using requests and BeautifulSoup.